chore(workspace): restore zero pedantic-clippy baseline
Clippy lint set has expanded since the v0.3.3 ship (rustc/clippy upgrade),
introducing 77 pedantic warnings across the workspace on main. This
commit restores the documented "zero pedantic warnings" invariant the
workspace enforces as of 2026-04-07.

Fixes applied:
- cargo clippy --fix autocorrections: format-string inlining
  ("{}", x → "{x}"), missing backticks in rustdoc, sort_unstable for
  primitive-tuple slices, sort_unstable_by for f32 keys, intra-doc
  backticks in place of quotes, primitive-type-inference cleanups, and
  a few items-after-test-module block reorderings.
- Source-level manual fixes: HashMap<String,_>::get("x") instead of
  .get(&"x".to_string()) (20 test sites in linear_combination.rs);
  FromIterator::from_iter → .collect() (3 sites in span.rs);
  redundant parens in finset.rs; loop-variable-as-index replaced with
  .iter().enumerate().take(n) in finset.rs; missing # Panics on
  utils::rand_perm; `impl Display` for CurvatureFoliation moved above
  #[cfg(test)] block.

Targeted #[allow(...)] annotations where the warning flags a legitimate
mathematical/testing pattern:
- clippy::float_cmp in gauge.rs tests (exact 0.0/1.0 from constructors)
  and e2_operad.rs test mod (identity disk has exact unit radius).
- clippy::similar_names in parallel-vs-sequential equivalence tests
  (seq_result / par_result pairs are the point), cap/cup and
  unit/comult mathematical pairings in compact_closed, triangle
  inequality variables in wasserstein.
- clippy::cast_possible_{truncation,wrap} in rayon_parallel and
  rayon_equivalence test fixtures (small sizes fit in i32 by
  construction).
- clippy::doc_link_with_quotes in mutation_workflows and
  composition_laws (single-quoted chars in docs are char literals,
  not intra-doc links).
- clippy::upper_case_acronyms on a local test type alias (UFM).
- clippy::needless_pass_by_value on make_branchial helper where the
  12+ call sites pass literal vec![] to avoid .clone() boilerplate.
- clippy::assertions_on_constants in gauge.rs tests (asserting the
  compile-time Lie-algebra axioms is the test's purpose).

Workspace test count unchanged at 918 passing / 0 failed / 0 ignored.

+++ catgraph-applied/tests/operad_boundary.rs
@@ -411,7 +411,7 @@ fn e1_go_to_monoid_reduces_to_single() {
 
     let product: f64 = e1.go_to_monoid(|(a, b)| f64::from(b - a));
 
-    let expected = f64::from(0.2) * f64::from(0.3) * f64::from(0.2);
+    let expected = 0.2 * 0.3 * 0.2;
     assert!(
         (product - expected).abs() < 1e-6,
         "Expected product ~{expected}, got {product}"
@@ -424,7 +424,7 @@ fn e1_go_to_monoid_single_interval() {
     let mut e1 = E1::new(vec![(0.1, 0.9)], true).unwrap();
     let result: f64 = e1.go_to_monoid(|(a, b)| f64::from(b - a));
 
-    let expected = f64::from(0.8);
+    let expected = 0.8;
     assert!(
         (result - expected).abs() < 1e-6,
         "Single interval monoid should be ~{expected}, got {result}"
